# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakPro API",
        "description": "API untuk aplikasi Ternak yang menyediakan berbagai fitur terkait ternak dan manajemen pengguna.",
        "contact": {
            "email": "dopidev@gmail.com"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/harga-pasar": {
            "get": {
                "tags": [
                    "Harga Pasar"
                ],
                "summary": "Mengambil semua data harga pasar",
                "description": "Endpoint ini digunakan untuk mengambil semua data harga pasar yang tersedia.",
                "operationId": "f91fc2e4083e922abc8ae81b4958c29f",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data harga pasar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HargaPasarItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Harga Pasar"
                ],
                "summary": "Menyimpan harga pasar baru",
                "description": "Endpoint ini digunakan untuk membuat data harga pasar baru.",
                "operationId": "5100d0bf973454c2a8cdc084bed9a56f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image_url",
                                    "nama",
                                    "harga_kg",
                                    "kondisi",
                                    "lokasi"
                                ],
                                "properties": {
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Sapi Holstein"
                                    },
                                    "harga_kg": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "kondisi": {
                                        "type": "string",
                                        "example": "Sehat"
                                    },
                                    "lokasi": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Harga pasar berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HargaPasarItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/harga-pasar/{id}": {
            "put": {
                "tags": [
                    "Harga Pasar"
                ],
                "summary": "Mengupdate data harga pasar",
                "description": "Endpoint ini digunakan untuk mengupdate data harga pasar berdasarkan ID.",
                "operationId": "060a9c954d816d02b4763620c2955d63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID harga pasar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Sapi Holstein"
                                    },
                                    "harga_kg": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "kondisi": {
                                        "type": "string",
                                        "example": "Sehat"
                                    },
                                    "lokasi": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Harga pasar berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HargaPasarItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Harga pasar tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Harga Pasar tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Harga Pasar"
                ],
                "summary": "Menghapus data harga pasar",
                "description": "Endpoint ini digunakan untuk menghapus data harga pasar berdasarkan ID.",
                "operationId": "ec725998f4d95a39635ebb893f153a7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID harga pasar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Harga pasar berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Harga Pasar berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Harga pasar tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Harga Pasar tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HargaPasarItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/image.jpg"
                    },
                    "nama": {
                        "type": "string",
                        "example": "Sapi Holstein"
                    },
                    "harga_kg": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "kondisi": {
                        "type": "string",
                        "example": "Sehat"
                    },
                    "lokasi": {
                        "type": "string",
                        "example": "Jakarta"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dalam format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}